Added more dynamic text descriptions.

+++ app/src/main/java/com/example/kotlinapp/ForecastDaily.kt
@@ -1,6 +1,6 @@
 package com.example.kotlinapp
 
 data class ForecastDaily(
-    val temp: Double,
+    val temp: Float,
     val description: String
 )
